% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{e_highlights}
\alias{e_highlights}
\alias{whe_highlights}
\alias{whe_highlights.data.frame}
\alias{whe_highlights.character}
\title{Get highlighted terms}
\usage{
whe_highlights(wh, flatten = FALSE)

\method{whe_highlights}{data.frame}(wh, flatten = FALSE)

\method{whe_highlights}{character}(wh, flatten = FALSE)
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \link[webhoser]{wh_collect}, see examples.}

\item{flatten}{whether to flatten ngrams.}
}
\description{
Extract highlighted terms
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"R programming language"') \%>\%  # use highlight!
  wh_collect() -> rstats # collect results

rstats \%>\%
  whe_highlights() -> highlights

highlights <- rstats \%>\%
  dplyr::mutate(
    nparagraphs = whe_highlights(text)
  )
}

}
