% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topmedia.R
\docType{data}
\name{topmedia}
\alias{topmedia}
\title{Top 100 media.}
\format{A data frame with 100 rows and 2 variables:
\describe{
  \item{name}{name of media}
  \item{thread.site}{site base url}
}}
\source{
World Economic Forum
}
\usage{
topmedia
}
\description{
A dataset containing top 100 media sources.
}
\keyword{datasets}
