% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{whe_date}
\alias{whe_date}
\title{Format date}
\usage{
whe_date(date)
}
\arguments{
\item{date}{date column, see examples.}
}
\value{
vector of class \code{POSIXct}.
}
\description{
Format dates returned by API, see details.
}
\details{
Webhose's servers are in Israel, this is used as timezone.
}
\examples{
\dontrun{
token <- wh_token("xXX-x0X0xX0X-00X")

rstats <- wh_news(token, q = '"R programming language"') \%>\%  # use quote marks!
  wh_collect() \%>\% # collect results
  dplyr::mutate(published = wh_date(published))
}

}
