% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexdiv.R
\name{whe_lexdiv}
\alias{whe_lexdiv}
\alias{whe_lexdiv.data.frame}
\alias{whe_lexdiv.character}
\title{Lexical diversity}
\usage{
whe_lexdiv(wh, measure = "R")

\method{whe_lexdiv}{data.frame}(wh, measure = "R")

\method{whe_lexdiv}{character}(wh, measure = "R")
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \link[webhoser]{wh_collect}, see examples.}

\item{measure}{lexical diversity type to compute, passed to \link[quanteda]{textstat_lexdiv}.}
}
\value{
if \code{data.frame} is passed will append a column named \code{lexdiv},
boolean.
If \code{character} vector is passed the function returns a \code{boolean} vector.
}
\description{
Calculate the lexical diversity or complexity of text.
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"R programming"') \%>\%
  wh_collect() -> rstats # collect results

rstats \%>\%
  whe_lexdiv() -> lexdiv

library(dplyr)

rstats \%>\%
  mutate(lexdiv = whe_lexdiv(text)) # pass text column
}

}
