% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentions_paragraph.R
\name{whe_mentioned}
\alias{whe_mentioned}
\alias{whe_mentioned.data.frame}
\alias{whe_mentioned.character}
\title{Get number of paragraphs with mentions}
\usage{
whe_mentioned(wh)

\method{whe_mentioned}{data.frame}(wh)

\method{whe_mentioned}{character}(wh)
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \link[webhoser]{wh_collect}, see examples.}
}
\value{
if \code{data.frame} is passed will append a column named \code{mentioned}.
If \code{character} vector is passed the function returns a \code{character} vector.
}
\description{
Counts number of paragraphs with mentions.
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = 'R OR Python', highlight = TRUE) \%>\%  # use highlight!
  wh_collect() -> rstats # collect results

rstats \%>\%
  whe_mentioned() -> mentioned

mentioned <- rstats \%>\%
  dplyr::mutate(nmentioned = whe_mentioned(highlightText))
}

}
