% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentions_paragraph.R
\name{whe_mentioned_wef}
\alias{whe_mentioned_wef}
\alias{whe_mentioned_wef.data.frame}
\alias{whe_mentioned_wef.character}
\title{Get number of paragraphs with mentions of the WEF}
\usage{
whe_mentioned_wef(wh)

\method{whe_mentioned_wef}{data.frame}(wh)

\method{whe_mentioned_wef}{character}(wh)
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \code{wh_collect}, see examples.}
}
\value{
if \code{data.frame} is passed will append a column named \code{mentioned.wef}.
If \code{character} vector is passed the function returns a \code{character} vector.
}
\description{
Counts number of paragraphs with mentions of the \code{World Economic Forum} or \code{WEF}.
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"World Economic Forum"', highlight = TRUE) \%>\%  # use highlight!
  wh_collect() -> wef # collect results

wef \%>\%
  whe_mentioned_wef() -> mentioned

mentioned <- wef \%>\%
  dplyr::mutate(nmentioned_wef = whe_mentioned_wef(highlightText))
}

}
\seealso{
\code{\link{whe_mentioned}}
}
