% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentions_paragraph.R
\name{whe_mentioned_wef_1p}
\alias{whe_mentioned_wef_1p}
\alias{whe_mentioned_wef_1p.data.frame}
\alias{whe_mentioned_wef_1p.character}
\title{Identify articles with WEF mentioned in 1st paragraph}
\usage{
whe_mentioned_wef_1p(wh)

\method{whe_mentioned_wef_1p}{data.frame}(wh)

\method{whe_mentioned_wef_1p}{character}(wh)
}
\arguments{
\item{wh}{\emph{text} object returned by \code{wh_collect}, see examples.}
}
\value{
if \code{data.frame} is passed will append a boolean column named \code{mentioned.wef.1p}.
If \code{character} vector is passed the function returns a \code{logical} vector.
}
\description{
Identifies if \code{World Economic Forum} or \code{WEF} is mentioned in 1st paragraph of article.
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"World Economic Forum"', highlight = TRUE) \%>\%  # use highlight!
  wh_collect() -> wef # collect results

wef \%>\%
  whe_mentioned_wef_1p() -> mentioned

mentioned <- wef \%>\%
  dplyr::mutate(nmentioned_wef = whe_mentioned_wef_1p(text))
}

}
\seealso{
\code{\link{whe_mentioned}}
}
