% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentions.R
\name{whe_mentions}
\alias{whe_mentions}
\alias{whe_mentions.data.frame}
\alias{whe_mentions.character}
\title{Get number of mentions}
\usage{
whe_mentions(wh)

\method{whe_mentions}{data.frame}(wh)

\method{whe_mentions}{character}(wh)
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \link[webhoser]{wh_collect}, see examples.}
}
\value{
if \code{data.frame} is passed will append a column named \code{nmentions}.
If \code{character} vector is passed the function returns a \code{character} vector.
}
\description{
Counts number of mentions in highlighted text.
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"R AND Python"') \%>\%  # use highlight!
  wh_collect() -> rstats # collect results

rstats \%>\%
  whe_mentions() -> nmentions

library(dplyr)
rstats \%>\%
  mutate(nmentions = whe_mentions(text)) -> mentions
}

}
