% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paragrahs.R
\name{whe_paragraphs}
\alias{whe_paragraphs}
\alias{whe_paragraphs.data.frame}
\alias{whe_paragraphs.character}
\title{Get number of paragraphs}
\usage{
whe_paragraphs(wh)

\method{whe_paragraphs}{data.frame}(wh)

\method{whe_paragraphs}{character}(wh)
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \link[webhoser]{wh_collect}, see examples.}
}
\value{
if \code{data.frame} is passed will append a column named \code{nparagraphs}.
If \code{character} vector is passed the function returns a \code{character} vector.
}
\description{
Counts number of paragraphs in text.
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"R programming" OR Python') \%>\%  # use highlight!
  wh_collect() -> rstats # collect results

rstats \%>\%
  whe_paragraphs() -> nparagraphs

library(dplyr)

rstats \%>\%
  mutate(nparagraphs = whe_paragraphs(text)) -> paragraphs
}

}
