% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram.R
\name{whe_post_gram}
\alias{whe_post_gram}
\alias{whe_post_gram.data.frame}
\alias{whe_post_gram.character}
\title{Get following bigrams}
\usage{
whe_post_gram(wh, n = 2, flatten = FALSE)

\method{whe_post_gram}{data.frame}(wh, n = 2, flatten = FALSE)

\method{whe_post_gram}{character}(wh, n = 2, flatten = FALSE)
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \link[webhoser]{wh_collect}, see examples.}

\item{n}{number of terms, defaults to 2.}

\item{flatten}{whether to flatten ngrams.}
}
\value{
if \code{data.frame} is passed will append a column named \code{pre.gram.n} where \code{n} is
the number of terms, parameter \code{n}. If \code{character} vector is passed will return list
(or vector dependeing on \code{flatten}) of n-grams.
}
\description{
Get following bigrams
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"World Economic Forum"') \%>\%  # use highlight!
  wh_collect() -> wef # collect results

wef \%>\%
  whe_post_gram(3) -> post_trigrams

library(dplyr)
wef \%>\%
  mutate(post.bigram = whe_post_gram(highlightText, flatten = TRUE)) -> post_bigrams
}

}
\seealso{
\code{\link{whe_pre_gram}}
}
