% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotations.R
\name{whe_quotes}
\alias{whe_quotes}
\alias{whe_quotes.data.frame}
\alias{whe_quotes.character}
\title{Identify quotations}
\usage{
whe_quotes(wh)

\method{whe_quotes}{data.frame}(wh)

\method{whe_quotes}{character}(wh)
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \link[webhoser]{wh_collect}, see examples.}
}
\value{
if \code{data.frame} is passed will append a column named \code{quotes},
boolean.
If \code{character} vector is passed the function returns a \code{boolean} vector.
}
\description{
Identify articles with quotations.
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"R programming language"') \%>\%
  wh_collect() -> rstats # collect results

rstats \%>\%
  whe_quotes() -> q

library(dplyr)

rstats \%>\%
  mutate(quotes = whe_quotes(text)) # pass text column
}

}
