% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{whe_search}
\alias{whe_search}
\alias{whe_search.data.frame}
\title{Search}
\usage{
whe_search(wh, search, output)

\method{whe_search}{data.frame}(wh, search, output, where = "text")
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \link[webhoser]{wh_collect}, see examples.}

\item{search}{regex to search, passed to \code{\link{grepl}}.}

\item{output}{name of output column.}

\item{where}{column where to search.}
}
\value{
Returns the number of times the \code{search} was found.
}
\description{
Search for terms.
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"R programming"') \%>\%  # use highlight!
  wh_collect() -> rstats # collect results

rstats \%>\%
  whe_search("Hadley|Wickham", "HadleyWickham") -> hadley
}

}
