% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{whe_search_1p}
\alias{whe_search_1p}
\alias{whe_search_1p.data.frame}
\title{Search in the first paragraph}
\usage{
whe_search_1p(wh, search, output)

\method{whe_search_1p}{data.frame}(wh, search, output)
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \link[webhoser]{wh_collect}, see examples.}

\item{search}{regex to search, passed to \code{\link{grepl}}.}

\item{output}{name of output column.}
}
\value{
if \code{data.frame} is passed will append a boolean column named \code{output}.
If \code{character} vector is passed the function returns a \code{logical} vector.
}
\description{
Search the number of times a regex is present in the first parargraph.
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"R programming" OR rstats', highlight = TRUE) \%>\%
  wh_collect() -> rstats # collect results

rstats \%>\%
  whe_search_1p("Hadley", "hadley")
}

}
\seealso{
\code{\link{whe_mentioned}}
}
