% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentences.R
\name{whe_sentences}
\alias{whe_sentences}
\alias{whe_sentences.data.frame}
\alias{whe_sentences.character}
\title{Get number of sentences}
\usage{
whe_sentences(wh)

\method{whe_sentences}{data.frame}(wh)

\method{whe_sentences}{character}(wh)
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \link[webhoser]{wh_collect}, see examples.}
}
\value{
if \code{data.frame} is passed will append a column named \code{nsentences}.
If \code{character} vector is passed the function returns a \code{character} vector.
}
\description{
Counts number of sentences in text.
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"R programming"') \%>\%
  wh_collect() -> rstats # collect results

rstats \%>\%
  whe_sentences() -> nsentences

library(dplyr)
rstats \%>\%
  mutate(nsentences = whe_sentences(text)) -> nsentences
}

}
