% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{whe_sentiment}
\alias{whe_sentiment}
\alias{whe_sentiment.character}
\alias{whe_sentiment.data.frame}
\title{Sentiment}
\usage{
whe_sentiment(wh, method = "afinn", ...)

\method{whe_sentiment}{character}(wh, method = "afinn", ...)

\method{whe_sentiment}{data.frame}(wh, method = "afinn", ...)
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \link[webhoser]{wh_collect}, see examples.}

\item{method}{Sentiment method passed to \link[syuzhet]{get_sentiment}.}

\item{...}{Any other parameters to pass to \link[syuzhet]{get_sentiment}.}
}
\description{
Get Sentiment score from text.
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"R programming language"') \%>\%
  wh_collect() -> rstats # collect results

sent <- rstats \%>\%
  whe_sentiment()

library(dplyr)

rstats \%>\%
  mutate(sent = whe_sentiment(text))
}

}
