% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{whe_similarity}
\alias{whe_similarity}
\alias{whe_similarity.data.frame}
\alias{whe_similarity.character}
\title{Compare articles with set of documents}
\usage{
whe_similarity(wh, docs, progress = interactive())

\method{whe_similarity}{data.frame}(wh, docs, progress = interactive())

\method{whe_similarity}{character}(wh, docs, progress = interactive())
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \link[webhoser]{wh_collect}, see examples.}

\item{docs}{documents to compare the articles with.}

\item{progress}{whether to show progress bar.}
}
\value{
if a \code{data.frame} is passed will append a column named \code{similarity.*} where \code{*} is the input document number.
If a \code{character} vector is passed the function returns a \code{character} vector.
}
\description{
Compare articles with set of documents, i.e.: press releases.
}
\details{
This function uses the \href{Jacquard's index}{https://en.wikipedia.org/wiki/Jaccard_index}
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"World Economic Forum"') \%>\%  # use highlight!
  wh_collect() -> wef # collect results

library(rvest)

html <- read_html('http://reports.weforum.org/global-gender-gap-report-2017/press-release/')

# scrape Gender Gap Report press release
html \%>\%
  html_nodes(".content") \%>\%
  html_children() \%>\%
  html_text() \%>\%
  .[5:40] \%>\%
  paste0(., collapse = "\\n") -> pr

wef \%>\%
  whe_similarity(pr) -> similarity

library(dplyr)
wef \%>\%
  mutate(nmentions = whe_mentions(text)) -> similarity
}

}
