% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topmedia.R
\name{whe_topmedia}
\alias{whe_topmedia}
\alias{whe_topmedia.data.frame}
\title{Identify Media}
\usage{
whe_topmedia(wh)

\method{whe_topmedia}{data.frame}(wh)
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \link[webhoser]{wh_collect}, see examples.}
}
\value{
if \code{data.frame} is passed will append a boolean column named \code{topmedia}.
}
\description{
Identify mentions from top media.
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"R programming"') \%>\%
  wh_collect() -> rstats # collect results

rstats \%>\%
  whe_topmedia() -> topmedia
}

}
