% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words.R
\name{whe_words}
\alias{whe_words}
\alias{whe_words.data.frame}
\alias{whe_words.character}
\title{Get number of words}
\usage{
whe_words(wh)

\method{whe_words}{data.frame}(wh)

\method{whe_words}{character}(wh)
}
\arguments{
\item{wh}{\emph{highlighted} object returned by \link[webhoser]{wh_collect}, see examples.}
}
\value{
if \code{data.frame} is passed will append a column named \code{nwords}.
If \code{character} vector is passed the function returns a \code{character} vector.
}
\description{
Counts number of words in text.
}
\examples{
\dontrun{
library(webhose)
token <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = '"R programming"') \%>\%
  wh_collect() -> rstats # collect results

rstats \%>\%
  whe_words() -> nwords

library(dplyr)
rstats \%>\%
  mutate(nwords = whe_words(text)) -> words
}

}
